+++ src/main.cpp
@@ -3,6 +3,7 @@
 
 // Qt import 
 #include <QApplication>
+#include <QFile>
 
 // App import
 #include "markdown_mainwindow.hpp"
@@ -10,12 +11,18 @@
 
 int main(int argc, char* argv[]) {
 
+    QFile style(":/style/qt-dark.qss");
+
     QApplication app(argc, argv);
+
+    if( style.open(QIODevice::ReadOnly | QIODevice::Text) )
+        app.setStyleSheet(style.readAll());
+
+    style.close();
 
     MarkdownEditor editor;
 
     editor.showMaximized();
 
     return app.exec();
-
 }