+++ src/markdown_webengine.cpp
@@ -1,5 +1,0 @@
-#include "markdown_webengine.hpp"
-
-MarkdownWebEngine::MarkdownWebEngine(QWidget* parent) : QWebEngineView(parent) {
-
-}